Refactored Arrange to operate by rows
